Generate static tag pages and simplify sidebar
