--- docs/assets/css/tags-ui.css
@@ -1,39 +1,0 @@
-/* Sidebar tag list */
-.tag-browser-header {
-  font-weight: 600;
-  margin: 0 0 0.5rem 0;
-}
-.tag-list {
-  list-style: none;
-  padding-left: 0;
-  margin: 0;
-}
-.tag-list li {
-  margin: 0.2rem 0;
-}
-.tag-link {
-  text-decoration: none;
-}
-
-/* Tags hub controls/results */
-.tag-hub {
-  margin: 1rem 0;
-}
-.tag-controls {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(180px, 1fr));
-  gap: 0.75rem;
-  align-items: center;
-  margin-bottom: 1rem;
-}
-.tag-results {
-  list-style: none;
-  padding-left: 0;
-}
-.tag-results li {
-  margin: 0.4rem 0;
-}
-.tag-results li small {
-  opacity: 0.7;
-  margin-left: 0.35rem;
-}
