+++ docs/tags/Bayesian.md
@@ -1,0 +1,10 @@
+---
+title: Bayesian
+hide:
+  - toc
+---
+
+# Bayesian
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
