+++ docs/tags/GLMM.md
@@ -1,0 +1,10 @@
+---
+title: GLMM
+hide:
+  - toc
+---
+
+# GLMM
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
