+++ docs/tags/INLA.md
@@ -1,0 +1,10 @@
+---
+title: INLA
+hide:
+  - toc
+---
+
+# INLA
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
