+++ docs/tags/LGM.md
@@ -1,0 +1,10 @@
+---
+title: LGM
+hide:
+  - toc
+---
+
+# LGM
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
