+++ docs/tags/R.md
@@ -1,0 +1,10 @@
+---
+title: R
+hide:
+  - toc
+---
+
+# R
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
