+++ docs/tags/analytics.md
@@ -1,0 +1,12 @@
+---
+title: analytics
+hide:
+  - toc
+---
+
+# analytics
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
+- [example-workflow](/analytics-library/example-workflow.md)  
+  <small></small>
