+++ docs/tags/biomass.md
@@ -1,0 +1,10 @@
+---
+title: biomass
+hide:
+  - toc
+---
+
+# biomass
+
+- [gedi](/data-library/gedi.md)  
+  <small></small>
