+++ docs/tags/burn-severity.md
@@ -1,0 +1,10 @@
+---
+title: burn-severity
+hide:
+  - toc
+---
+
+# burn-severity
+
+- [fire-cbi](/data-library/fire-cbi.md)  
+  <small></small>
