+++ docs/tags/climate.md
@@ -1,0 +1,10 @@
+---
+title: climate
+hide:
+  - toc
+---
+
+# climate
+
+- [drought](/data-library/drought.md)  
+  <small></small>
