+++ docs/tags/contribution.md
@@ -1,0 +1,10 @@
+---
+title: contribution
+hide:
+  - toc
+---
+
+# contribution
+
+- [how-to-contribute](/quickstart/data-library/how-to-contribute.md)  
+  <small></small>
