+++ docs/tags/data library.md
@@ -1,0 +1,12 @@
+---
+title: data library
+hide:
+  - toc
+---
+
+# data library
+
+- [how-to-use](/quickstart/data-library/how-to-use.md)  
+  <small></small>
+- [how-to-contribute](/quickstart/data-library/how-to-contribute.md)  
+  <small></small>
