+++ docs/tags/data-management.md
@@ -1,0 +1,10 @@
+---
+title: data-management
+hide:
+  - toc
+---
+
+# data-management
+
+- [move-data-to-instance](/data-library/move-data-to-instance.md)  
+  <small></small>
