+++ docs/tags/development.md
@@ -1,0 +1,10 @@
+---
+title: development
+hide:
+  - toc
+---
+
+# development
+
+- [dev-schedule](/dev-schedule.md)  
+  <small>2025-08-14</small>
