+++ docs/tags/drought.md
@@ -1,0 +1,12 @@
+---
+title: drought
+hide:
+  - toc
+---
+
+# drought
+
+- [disturbance-stack](/data-library/disturbance-stack.md)  
+  <small></small>
+- [drought](/data-library/drought.md)  
+  <small></small>
