+++ docs/tags/education.md
@@ -1,0 +1,10 @@
+---
+title: education
+hide:
+  - toc
+---
+
+# education
+
+- [advanced-textbook](/quickstart/advanced-textbook.md)  
+  <small></small>
