+++ docs/tags/example.md
@@ -1,0 +1,12 @@
+---
+title: example
+hide:
+  - toc
+---
+
+# example
+
+- [example-container](/container-library/example-container.md)  
+  <small></small>
+- [example-workflow](/analytics-library/example-workflow.md)  
+  <small></small>
