+++ docs/tags/fia.md
@@ -1,0 +1,10 @@
+---
+title: fia
+hide:
+  - toc
+---
+
+# fia
+
+- [fia](/data-library/fia.md)  
+  <small></small>
