+++ docs/tags/fire.md
@@ -1,0 +1,12 @@
+---
+title: fire
+hide:
+  - toc
+---
+
+# fire
+
+- [fire-cbi](/data-library/fire-cbi.md)  
+  <small></small>
+- [landfire-events](/data-library/landfire-events.md)  
+  <small></small>
