+++ docs/tags/forest.md
@@ -1,0 +1,12 @@
+---
+title: forest
+hide:
+  - toc
+---
+
+# forest
+
+- [fia](/data-library/fia.md)  
+  <small></small>
+- [treemap](/data-library/treemap.md)  
+  <small></small>
