+++ docs/tags/gdal.md
@@ -1,0 +1,10 @@
+---
+title: gdal
+hide:
+  - toc
+---
+
+# gdal
+
+- [mounting-via-vsi](/data-library/mounting-via-vsi.md)  
+  <small></small>
