+++ docs/tags/gedi.md
@@ -1,0 +1,10 @@
+---
+title: gedi
+hide:
+  - toc
+---
+
+# gedi
+
+- [gedi](/data-library/gedi.md)  
+  <small></small>
