+++ docs/tags/inventory.md
@@ -1,0 +1,10 @@
+---
+title: inventory
+hide:
+  - toc
+---
+
+# inventory
+
+- [fia](/data-library/fia.md)  
+  <small></small>
