+++ docs/tags/landcover.md
@@ -1,0 +1,10 @@
+---
+title: landcover
+hide:
+  - toc
+---
+
+# landcover
+
+- [lcmap](/data-library/lcmap.md)  
+  <small></small>
