+++ docs/tags/landfire.md
@@ -1,0 +1,12 @@
+---
+title: landfire
+hide:
+  - toc
+---
+
+# landfire
+
+- [disturbance-stack](/data-library/disturbance-stack.md)  
+  <small></small>
+- [landfire-events](/data-library/landfire-events.md)  
+  <small></small>
