+++ docs/tags/lidar.md
@@ -1,0 +1,10 @@
+---
+title: lidar
+hide:
+  - toc
+---
+
+# lidar
+
+- [gedi](/data-library/gedi.md)  
+  <small></small>
