+++ docs/tags/oasis.md
@@ -1,0 +1,10 @@
+---
+title: oasis
+hide:
+  - toc
+---
+
+# oasis
+
+- [Starting with OASIS](/quickstart/oasis.md)  
+  <small>2024-01-01</small>
