+++ docs/tags/quickstart.md
@@ -1,0 +1,14 @@
+---
+title: quickstart
+hide:
+  - toc
+---
+
+# quickstart
+
+- [Starting with OASIS](/quickstart/oasis.md)  
+  <small>2024-01-01</small>
+- [how-to-use](/quickstart/data-library/how-to-use.md)  
+  <small></small>
+- [how-to-contribute](/quickstart/data-library/how-to-contribute.md)  
+  <small></small>
