+++ docs/tags/sentinel-2.md
@@ -1,0 +1,10 @@
+---
+title: sentinel-2
+hide:
+  - toc
+---
+
+# sentinel-2
+
+- [Pull_Sentinal2_l2_data](/data-library/Pull_Sentinal2_l2_data.md)  
+  <small></small>
