+++ docs/tags/spatial.md
@@ -1,0 +1,10 @@
+---
+title: spatial
+hide:
+  - toc
+---
+
+# spatial
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
