+++ docs/tags/spatiotemporal.md
@@ -1,0 +1,10 @@
+---
+title: spatiotemporal
+hide:
+  - toc
+---
+
+# spatiotemporal
+
+- [INLA — Drop-in Analytics Module](/analytics-library/inla.md)  
+  <small></small>
