+++ docs/tags/stac.md
@@ -1,0 +1,12 @@
+---
+title: stac
+hide:
+  - toc
+---
+
+# stac
+
+- [stac_simple](/data-library/stac_simple.md)  
+  <small></small>
+- [stac_mount_save](/data-library/stac_mount_save.md)  
+  <small></small>
