+++ docs/tags/treemap.md
@@ -1,0 +1,10 @@
+---
+title: treemap
+hide:
+  - toc
+---
+
+# treemap
+
+- [treemap](/data-library/treemap.md)  
+  <small></small>
