+++ docs/tags/tutorial.md
@@ -1,0 +1,16 @@
+---
+title: tutorial
+hide:
+  - toc
+---
+
+# tutorial
+
+- [move-data-to-instance](/data-library/move-data-to-instance.md)  
+  <small></small>
+- [stac_simple](/data-library/stac_simple.md)  
+  <small></small>
+- [stac_mount_save](/data-library/stac_mount_save.md)  
+  <small></small>
+- [mounting-via-vsi](/data-library/mounting-via-vsi.md)  
+  <small></small>
