+++ docs/tags/usage.md
@@ -1,0 +1,10 @@
+---
+title: usage
+hide:
+  - toc
+---
+
+# usage
+
+- [how-to-use](/quickstart/data-library/how-to-use.md)  
+  <small></small>
